@@ -1,3 +1,5 @@
+'use client';
+
 import cn from 'classnames';
 import React, { forwardRef, useRef, ButtonHTMLAttributes } from 'react';
 import { mergeRefs } from 'react-merge-refs';
@@ -52,7 +54,7 @@ const Button = forwardRef<HTMLButtonElement, Props>((props, buttonRef) => {
     >
       {children}
       {loading && (
-        <i className="pl-2 m-0 flex">
+        <i className="flex pl-2 m-0">
           <LoadingDots />
         </i>
       )}
