@@ -3,29 +3,27 @@ import Link from 'next/link';
 import Logo from '@/components/icons/Logo';
 import GitHub from '@/components/icons/GitHub';
 
-import s from './Footer.module.css';
-
 export default function Footer() {
   return (
     <footer className="mx-auto max-w-[1920px] px-6 bg-zinc-900">
-      <div className="grid grid-cols-1 lg:grid-cols-12 gap-8 border-b border-zinc-600 py-12 text-white transition-colors duration-150 bg-zinc-900">
+      <div className="grid grid-cols-1 gap-8 py-12 text-white transition-colors duration-150 border-b lg:grid-cols-12 border-zinc-600 bg-zinc-900">
         <div className="col-span-1 lg:col-span-2">
           <Link
             href="/"
-            className="flex flex-initial items-center font-bold md:mr-24"
+            className="flex items-center flex-initial font-bold md:mr-24"
           >
-            <span className="rounded-full border border-zinc-700 mr-2">
+            <span className="mr-2 border rounded-full border-zinc-700">
               <Logo />
             </span>
             <span>ACME</span>
           </Link>
         </div>
         <div className="col-span-1 lg:col-span-2">
-          <ul className="flex flex-initial flex-col md:flex-1">
+          <ul className="flex flex-col flex-initial md:flex-1">
             <li className="py-3 md:py-0 md:pb-4">
               <Link
                 href="/"
-                className="text-white hover:text-zinc-200 transition ease-in-out duration-150"
+                className="text-white transition duration-150 ease-in-out hover:text-zinc-200"
               >
                 Home
               </Link>
@@ -33,7 +31,7 @@ export default function Footer() {
             <li className="py-3 md:py-0 md:pb-4">
               <Link
                 href="/"
-                className="text-white hover:text-zinc-200 transition ease-in-out duration-150"
+                className="text-white transition duration-150 ease-in-out hover:text-zinc-200"
               >
                 About
               </Link>
@@ -41,7 +39,7 @@ export default function Footer() {
             <li className="py-3 md:py-0 md:pb-4">
               <Link
                 href="/"
-                className="text-white hover:text-zinc-200 transition ease-in-out duration-150"
+                className="text-white transition duration-150 ease-in-out hover:text-zinc-200"
               >
                 Careers
               </Link>
@@ -49,7 +47,7 @@ export default function Footer() {
             <li className="py-3 md:py-0 md:pb-4">
               <Link
                 href="/"
-                className="text-white hover:text-zinc-200 transition ease-in-out duration-150"
+                className="text-white transition duration-150 ease-in-out hover:text-zinc-200"
               >
                 Blog
               </Link>
@@ -57,16 +55,16 @@ export default function Footer() {
           </ul>
         </div>
         <div className="col-span-1 lg:col-span-2">
-          <ul className="flex flex-initial flex-col md:flex-1">
+          <ul className="flex flex-col flex-initial md:flex-1">
             <li className="py-3 md:py-0 md:pb-4">
-              <p className="text-white font-bold hover:text-zinc-200 transition ease-in-out duration-150">
+              <p className="font-bold text-white transition duration-150 ease-in-out hover:text-zinc-200">
                 LEGAL
               </p>
             </li>
             <li className="py-3 md:py-0 md:pb-4">
               <Link
                 href="/"
-                className="text-white hover:text-zinc-200 transition ease-in-out duration-150"
+                className="text-white transition duration-150 ease-in-out hover:text-zinc-200"
               >
                 Privacy Policy
               </Link>
@@ -74,15 +72,15 @@ export default function Footer() {
             <li className="py-3 md:py-0 md:pb-4">
               <Link
                 href="/"
-                className="text-white hover:text-zinc-200 transition ease-in-out duration-150"
+                className="text-white transition duration-150 ease-in-out hover:text-zinc-200"
               >
                 Terms of Use
               </Link>
             </li>
           </ul>
         </div>
-        <div className="col-span-1 lg:col-span-6 flex items-start lg:justify-end text-white">
-          <div className="flex space-x-6 items-center h-10">
+        <div className="flex items-start col-span-1 text-white lg:col-span-6 lg:justify-end">
+          <div className="flex items-center h-10 space-x-6">
             <a
               aria-label="Github Repository"
               href="https://github.com/vercel/nextjs-subscription-payments"
@@ -92,9 +90,11 @@ export default function Footer() {
           </div>
         </div>
       </div>
-      <div className="py-12 flex flex-col md:flex-row justify-between items-center space-y-4 bg-zinc-900">
+      <div className="flex flex-col items-center justify-between py-12 space-y-4 md:flex-row bg-zinc-900">
         <div>
-          <span>&copy; 2020 ACME, Inc. All rights reserved.</span>
+          <span>
+            &copy; {new Date().getFullYear()} ACME, Inc. All rights reserved.
+          </span>
         </div>
         <div className="flex items-center">
           <span className="text-white">Crafted by</span>
