@@ -1,49 +1,42 @@
 import Link from 'next/link';
-import { useRouter } from 'next/router';
-import { useSupabaseClient } from '@supabase/auth-helpers-react';
+import { createServerSupabaseClient } from '@/app/supabase-server';
 
 import Logo from '@/components/icons/Logo';
-import { useUser } from '@/utils/useUser';
+import SignOutButton from './SignOutButton';
 
 import s from './Navbar.module.css';
 
-const Navbar = () => {
-  const router = useRouter();
-  const supabaseClient = useSupabaseClient();
-  const { user } = useUser();
+export default async function Navbar() {
+  const supabase = createServerSupabaseClient();
+  const {
+    data: { user }
+  } = await supabase.auth.getUser();
 
   return (
     <nav className={s.root}>
       <a href="#skip" className="sr-only focus:not-sr-only">
         Skip to content
       </a>
-      <div className="mx-auto max-w-6xl px-6">
-        <div className="flex justify-between align-center flex-row py-4 md:py-6 relative">
-          <div className="flex flex-1 items-center">
+      <div className="max-w-6xl px-6 mx-auto">
+        <div className="relative flex flex-row justify-between py-4 align-center md:py-6">
+          <div className="flex items-center flex-1">
             <Link href="/" className={s.logo} aria-label="Logo">
               <Logo />
             </Link>
-            <nav className="space-x-2 ml-6 hidden lg:block">
+            <nav className="hidden ml-6 space-x-2 lg:block">
               <Link href="/" className={s.link}>
                 Pricing
               </Link>
-              <Link href="/account" className={s.link}>
-                Account
-              </Link>
+              {user && (
+                <Link href="/account" className={s.link}>
+                  Account
+                </Link>
+              )}
             </nav>
           </div>
-
-          <div className="flex flex-1 justify-end space-x-8">
+          <div className="flex justify-end flex-1 space-x-8">
             {user ? (
-              <span
-                className={s.link}
-                onClick={async () => {
-                  await supabaseClient.auth.signOut();
-                  router.push('/signin');
-                }}
-              >
-                Sign out
-              </span>
+              <SignOutButton />
             ) : (
               <Link href="/signin" className={s.link}>
                 Sign in
@@ -54,6 +47,4 @@ const Navbar = () => {
       </div>
     </nav>
   );
-};
-
-export default Navbar;
+}
